selected an element and appended newly created list item
In the else block, this line selects an element with the ID list using jQuery ($('#list')) and appends the newly created list item (li) to it. It's equivalent to document.querySelector('#list').appendChild(li) in pure JavaScript. This assumes that there is an HTML element with the ID list where the list items will be appended.